Remove secret_storage::some_secret module.
